Enable alternate g++ versions in visionBuilder.
Confirming the existence of a serious compiler bug in g++ version 5 (the
default for ubuntu 16.04) required the ability to run alternative versions
of g++.  While not a problem in general on Linux, 'visionBuilder' and its
supporting scripts and make file templates required a few updates to pass
the required options into the 'make' command.  Specifically, alternative
installed versions of g++ can now be selected by setting the CXX
environment variable prior to running 'visionBuilder' or 'buildRelease'.

Along for the ride in this commit are minor adjustments to make options
generator in 'visionBuilder'.  Specifically:

1) the '-j' option that specifies parallel builds, when given without a
   numeric argument, automatically uses all available CPUs.  That was
   effectively our old default; however, we unnecessarily went about
   determining the CPU count with no option for override or the ability
   to make use of hot-plugged CPUs added during course of a build.  The
   use of the bare '-j' covers the latter case; the examination of the
   CPUCOUNT environment variable the former.
2) the '-k' option (continue building on errors) has long been our hidden
   default.  Make the specification of that option explicit in the
   subroutine that specifies all other 'make' options.
3) temporarily remove a stray 'CC -V' command (probably applicable on
   Solaris) that doesn't work on Linux.