Add class VdbNetwork to the V shared library.
VdbNetwork, a class that provides NDF access and locking services for
vision databases is currently used by two executables - batchvision
(the vision engine) and ndftool (a tool for examining and editing
transaction history).  On some platforms it is separately compiled
and linked to those executables, on others it is available in the V
(libV) shared library.  Standardize that by always making it part of
the shared library.