Add VkLongUnsigneds to the V shared library.
VkLongUnsigneds contains class and template definitions used by two
mainstream executables - batchvision and dbupdate - and one utility
executable - showsizes.  Prior to this commit, VkLongSigneds.cpp
was compiled and linked separately for each executable.  In keeping
with the need and desire for a single compiled instance for all
shared objects, add VkLongUnsigneds to the V shared library.