Disable atomic memory operations in Mac port for now.
Temporarily use the no-atomics implementation of atomic memory operations
in the Mac port to allow the port to focus on the big picture of the Mac
runtime and compilation environment.  A future commit will return to this
to make use of the Mac specific version of atomic memory operations (see
kern/OSAtomic.h).