Shared libraries are named differently on the Mac.
Shared (dynamic) libraries are named and work slightly differently
on the Mac.  To begin with, their names end in '.dylib', not '.so'
or '.so.#', so the first step in getting them to build on the Mac
is changing their 'visionBuilder' target names to remove those
platform dependent suffixes.  To keep 'git' happy, this commit
handles the file renaming part of a two part change.  The next
commit will actually make the required content changes.