Set Mac -install_name using a make recipe, not VLINK-CC
This commit undoes the hack of using VLINK-CC to set a hardwired '-install_name'
in our Mac libraries, introducing a new 'linkerrule' meta-tag to do the job
properly using 'make' (See any of the 'make.*.lib' files in this commit for an
example).

This commit also removes the 'make.rule.*' make recipe files for all of our
shared libraries.  With the introduction of 'linkerrule', the handling of target
directory creation in VLINK-CC where it belongs, and the changing of our
shared library suffix to 'so' instead of 'so.1' on Linux and Solaris, the
custom (and identical) recipes contained in those files are no long needed.