Working Mac implementation of our atomic memory operations.
This commit implements our atomic memory operations on the Mac using what
turns out to be the newly deprecated MAC <libkern/OSatomic.h> atomic API.
This is interesting because the declarations in this header have been
deprecated in favor of C++11's std::atomic (<atomic>) API.  While this is
more work than it's worth as part of this initial Mac port, it is worth it
longer term.  At the moment, <libkern/OSatomic.h> (man atomic_deprecated)
with its function call interface is a better match for our current trait-based
manipulation of un-encapsulated atomic POD values.  In contrast, the proposed
C++11 alternative uses classes to encapsulate atomic POD values, making it
more object oriented but less of a drop-in fit for our current implementation.
Still:

    - It's part of a platform and architecture independent standard that's
      already been implemented for Linux, the Mac, probably Windows, and,
      maybe even someday, Solaris.
    - We already encapsulate atomic values in classes that are a natural fit
      (VCounter, VReferenceable, VSpinlock, etc.) for the std::atomic API,
      making the effort largely one of careful update, not major rewrite.
    - we can get rid of the '-Wno-deprecated-declarations' compiler option
      from the Mac build :-)