The last of the de-K&R-ification of the frontend.
What remains is the harder stuff requiring rewritings that respect 'const'
qualifiers, almost exclusively of literal strings. Of use in taking it the
rest of the way is the fact that the C++ compiler was used to get this far,
so it's now possible to use classes from V to help out.  On the other hand,
it may still more tedious work than it's worth.  Stay tuned...