Disable handle cycle detection in the 8.1 garbage collector
Building a bootstrap database on the Mac uncovered a bug in the release
8.1 version of the garbage collector.  Note that this is a logic issue
and NOT a MacOS specific problem; however, it does not manifest itself
in similar fashion on those platforms.  In all likelihood that is the
result of timing differences in the re-use of reclaimed memory however
that hypothesis has not been confirmed.

In a nutshell, it appears that the 8.1 garbage collector is failing to
retain containers indirectly referenced by handle alone.  The following
structure exposes the problem:

   (Vector)
      |
      |     ... POP via USegStore
      |
      V
  (Closure)
      |
      |     ... Handle via Context
      |
      V
  (Context) ... only reference is from Closure handle
      |
      |     ... Handle via Descriptor Store
      |
      V
   (Store)  ... only reference is from Context handle

There are a couple of different things going on that jointly conspire
to cause a problem:

    - Handle based retention is based solely on logic in the 8.1 garbage
      collector that attempts to determine if a handle is referenced from
      'outside' the mesh of containers. There is no provision in the 8.1
      garbage collector to directly mark a container for retention based
      on references from other handles.
    - The discovery algorithm is essentially a one pass algorithm that
      relies on a seed set of container table entries reachable only by
      container handle.  At the conclusion of that algorithm's execution,
      any container (handle) that proves reachable exclusively from one
      of the handles in the initial seed set is considered unreachable
      and therefore reclaimable.  There is no provision for re-running the
      discovery algorithm excluding container table entries originally in
      the seed set that prove to be reachable based on the algorithm's result.

Without going into too much detail regarding the discovery algorithm,
here's what appears to be happening:

    - The POP reference from (Vector)->(Closure) keeps (Closure) alive
      and keeps (Closure) from appearing in the initial seed set.
    - Because there are no POP references to either (Context) or (Store),
      they are both candidates for inclusion in the discovery algorithm's
      seed set.
    - Because (Closure) is not in the seed set, the discovery algorithm
      correctly concludes that (Context) is externally referenced and must
      be preserved.  (Store), unfortunately, does not share that fate.
      Because (Context) was in the seed set, the discovery algorithm finds
      the one and only reference to (Store) 'inside' the container mesh.
      Based on that discovery, it concludes that (Store) can be reclaimed
      despite the fact that it has just proved that (Context) should not
      have been in the seed set in the first place.

In the short term, it is possible to 'fix' the 8.1 garbage collector by
returning to the rough assumption of the 8.0 collector that a container
handle automatically implies reachability.  That is what this commit
does; however, that is not a long term fix given the potential (albeit
rare at the moment) for handle-only cycles in 8.1.