Generalize 'visitReferencesUsing' visitor to be an abstract Visitor object
Prior to these changes, the visitor parameter of 'visitReferencesUsing' was a pointer
to a VContainerHandle member.  That parameter is now a pointer to an abstract Visitor
object, enabling the creation of visitors not restricted to processing just container
handles (CPDs come to mind, require visitor specific logic, and are definitely next).