Remove 'this' null test from M_ASD::GetCTE
This is probably the trickiest of the refactorings needed to eliminate null
testing 'this' pointers.  As is the case for all such refactorings, either
the null test must be proven unnecessary or the required test must be moved
to the callers of the routines from which the test must be eliminated.  In
the case of these changes, the test is needed and must be moved.  Fortunately,
the affected code paths in this case filtered through relatively few routines
that all used an implicit 'Maybe' functional implementation pattern.  Test
deck and bootstrap tests passed; however, more in-the-wild testing of this
particular set of changes is in order.