Fix improper use of static_cast revealed by node.js testing on Windows
Testing our node.js native module on Windows revealed a 'missed mis-use' of a
'static_cast<T>()'.  The 'mis-use' occurred in the reversal of one of the type
erasures used in the implementation of Vxa.  It was 'missed' when the invariant
upon which it relied was invalidated by ongoing Vxa development.

In as compact a nutshell as possible, here's what happened.

Vxa provides the machinery for making the methods of 'stylized' C++ classes
callable from Vision by providing:

1) a base class (Vxa::VCollectableObject) from which user classes defining
   exposed functionality are derived;
2) a collection (clustering) mechanism used to bridge the gap between C++'s
   data-serial and Vision's data-parallel execution models;
3) a meta-data capture and translation mechanism connecting items 1 and 2
   that makes extensive use of C++ meta-programming techniques including
   type erasure.

As originally designed, Vxa supported only a flat inheritance model for user
defined classes exposing their C++ methods to Vision:

                       derived_1 -------+
                                        |
                       derived_2 -------+---->  Vxa::VCollectableObject
                                        |
                       derived_3 -------+

That model was loosely mirrored on the clustering side by an equally flat
cluster type model:

    Vxa::V...CollectionOf<derived_1> ---+
                                        |
    Vxa::V...CollectionOf<derived_2> ---+---->  Vxa::VCollection
                                        |
    Vxa::V...CollectionOf<derived_3> ---+

In that context, cluster type was always uniquely determined by the user
defined type.  Given a method associated with user defined type 'T', the
cluster type was always and simply 'Vxa::VCollectableCollectableOf<T>'.
That 'fact' was used during the instantiation of Vxa tasks used to apply
(call) methods for a subset of the objects in a cluster:

    template <typename T>
    Vxa::VCollectableMethod<T>::Implementation::Implementation (
        VCallData const &rCallData
    ) : BaseClass (rCallData), m_pCluster (
        static_cast<cluster_t*>(rCallData.cluster ())
    ) {
    }

where 'rCallData.cluster()' returns an object of type 'Vxa::VCollection' and
'cluster_t' is the expected T dependent 'typedef'.

Along the way, the flat inheritance model for user defined classes was
relaxed:

    derived_1a ---+---> derived_1 ------+
                  |                     |
    derived_1b ---+                     |
                        derived_2 ------+----->  Vxa::VCollectableObject
                                        |
                                        |
    derived_3b -------> derived_3 ------+

For good reason, the flat cluster type model was preserved:

    Vxa::V...CollectionOf<derived_1>- --+
                                        |
    Vxa::V...CollectionOf<derived_1a> --+
                                        |
    Vxa::V...CollectionOf<derived_1b> --+
                                        |
    Vxa::V...CollectionOf<derived_2> ---+---->  Vxa::VCollection
                                        |
    Vxa::V...CollectionOf<derived_3> ---+
                                        |
    Vxa::V...CollectionOf<derived_3a> --+

While the 'static_cast' remains well-formed, a method defined, for example, in
'class derived_1' can be invoked on a cluster type paramaterized by 'derived_1'
or any of its subclasses.  In Windows testing, a segmentation fault occurred
in the case of subclasses.  Interestingly, the same problem did not occur in
Linux or Mac testing, perhaps because the shape of the code generated by the
g++ and clang compilers is the same for the cluster template instantiations.