Implement VSingleCoreExecution and VcaEnableBlockingSockets as options to fix github issues #60 and #61.
You have to call sched_setaffinity before the first libV managed thread is spawned.  Because sched_setaffinity modifies a thread specific attribute, not a process specific attribute.  From the docs:

"The affinity mask is actually a per-thread attribute that can be adjusted independently for each of the threads in a thread group. The value returned from a call to gettid(2) can be passed in the argument pid. Specifying pid as 0 will set the attribute for the calling thread, and passing the value returned from a call to getpid(2) will set the attribute for the main thread of the thread group."