Fix segmentation fault described in issue #70
Remote Vxa objects are proxied in Vision by Vision Object clusters.  Their
methods are considered hidden methods of their proxy Object clusters while
their identities are isomorphic to the identities of the instances present
in those clusters.  Internally, those identities are primitively available
as the value of '^current' (getCurrent ()).  VSNFTask.cpp asked for ^self
instead.  That will reliably fail to be correct when ^self is a extension
of ^current.  While external objects are not usually explicitly extended,
it is possible as issue #70 reveals.